+++ src/node-project.ts
@@ -49,9 +49,13 @@ export interface NodeProjectOptions extends ProjectOptions, NodePackageOptions {
   /**
    * The name of the main release branch.
    *
-   * @default "master"
+   * NOTE: this field is temporarily required as we migrate the default value
+   * from "master" to "main". Shortly, it will be made optional with "main" as
+   * the default.
+   *
+   * @default "main"
    */
-  readonly defaultReleaseBranch?: string;
+  readonly defaultReleaseBranch: string;
 
   /**
    * Define a GitHub workflow for building PRs.
@@ -75,7 +79,7 @@ export interface NodeProjectOptions extends ProjectOptions, NodePackageOptions {
   readonly codeCovTokenSecret?: string;
 
   /**
-   * Define a GitHub workflow for releasing from "master" when new versions are
+   * Define a GitHub workflow for releasing from "main" when new versions are
    * bumped. Requires that `version` will be undefined.
    *
    * @default - true if not a subproject
@@ -100,7 +104,7 @@ export interface NodeProjectOptions extends ProjectOptions, NodePackageOptions {
    *
    * Default value is based on defaultReleaseBranch.
    *
-   * @default [ "master" ]
+   * @default [ "main" ]
    */
   readonly releaseBranches?: string[];
 
@@ -288,7 +292,7 @@ export interface NodeProjectOptions extends ProjectOptions, NodePackageOptions {
  */
 export enum AutoRelease {
   /**
-   * Automatically bump & release a new version for every commit to "master"
+   * Automatically bump & release a new version for every commit to "main"
    */
   EVERY_COMMIT,
 
@@ -512,7 +516,11 @@ export class NodeProject extends Project {
       this.addDevDeps(`projen@${projenVersion}`);
     }
 
-    const defaultReleaseBranch = options.defaultReleaseBranch ?? 'master';
+    if (!options.defaultReleaseBranch) {
+      throw new Error('"defaultReleaseBranch" is temporarily a required option while we migrate its default value from "master" to "main"');
+    }
+
+    const defaultReleaseBranch = options.defaultReleaseBranch ?? 'main';
 
     // version is read from a committed file called version.json which is how we bump
     this._version = new Version(this, { releaseBranch: defaultReleaseBranch });